commands/show_info: fix formatting error for C method header
Fixes #2006 (show-source formatting bug in From)

In order to test this I had to build Pry from master locally and install
it. When I ran `String#initialize` I saw an error that `has_pry_doc` is
undefined on the Config class. This makes sense. This option is defined in the
pry-doc plugin.

As a workaround, I had to change the check to `defined?`. However, in the
future, we should make Pry plugin-agnostic - strictly no plugin checks in the
code.