config: delete the `control_d_handler` option
This change is a preparational step for #1824 (Enabling `frozen_string_literal:
true` in `~/.pryc` crashes most operations).

The handler was mutating `eval_string` that we pass, and this behaviour will
stop working in Ruby 3.

I was pondering about keeping this option but couldn't come up with an easy
solution. I made a GitHub search to see if anybody uses it, and it seems like
nobody configures it. I am not sure why anybody would want to do that, so I hope
it won't be missed.

The handler's behaviour wasn't deleted, it is still in place. We just moved the
handler's code to `pry_instance.rb`.