Resolve config warnings about redefined method
I spotted "method redefined" warnings while working with Pry in another project,
which used `--warnings`. It would be nice to enable `--warnings` in RSpec,
however it's a bit tricky because our test suite has a lot of violations (on
purpose). We should strive to fix all possible warnings, though (especially
those that occur in `lib/`).