spec/commands/show_{doc,source}: refactor to be order independent
A bunch of tests were order-dependent and weren't cleaning up properly after
they're executed. This results in complications with #2055 where we want to
randomise our test suite.

With this change every test defines what it needs and then cleans up properly,
instead of relying on the global space and fixtures.

P.S. The specs are still quite messy but this change is a step in the right
direction.