Directly delegate internally-used Method methods
Relying on `method_missing` for all delegation to `Method` methods means
that it's easier for the wrong method to be called when gems mess with
`Object` or other such core areas.

See https://github.com/ankane/ownership/pull/3 for an example of this.

By defining explicit delegators, at least for the methods that we use
internally withing Pry, we can eliminate this issue and be more
communicative around the methods on the `Pry::Method` class.

For `source_location`, I went with a direct method definition using
`__send__` due to `Forwardable` complaining that it was being told to
delegate to a private method. `Method#source_location` isn't private,
but `NilClass#source_location` is, so the direct definition alleviated the
warning message for when `@method` is undefined.