Fix whereami -c when class definition is not beginning of line
Consider the following class definition:
```
out = class Cor
  def blimey; end
  pry_eval(binding, 'whereami -c')
end
```

For classes defined as part of an expression like this, "whereami -c"
was failing to find the class definition. To fix this, we modify the
regular expression so the "class" keyword is no longer required to
appear at the beginning of the line IF there is an equals sign preceding
it.

https://github.com/pry/pry/issues/2052