exceptions: check if $SAFE is supported by Ruby
Fixes #2099 ($SAFE will become a normal global variable in Ruby 3.0)

MRI < 2.7 supports `$SAFE` variable, which is associated with
taint/untaint methods. All of that is deprecated in 2.7 with a lot of
warning messages generated, and no other effect.

None of this is supported on other Ruby platforms.

Related Ruby ticket:
https://bugs.ruby-lang.org/issues/8468