Improve help listing for regex aliases
This commit improves the appearance of regex aliases in the help index
by storing the result of calling #inspect on the regex as the listing.

For example, consider the `whereami` alias `/whereami[!?]+/`.
Previously this would appear in the help index as
`(?-mix:whereami[!?]+)`. This commit fixes this so it appears as
`/whereami[!?]+/`.