Restore `--no-history` CLI flag functionality
The `--no-history` option was almost ten years ago, in db649858a, when
the object at `Pry.config.history` was still an `OpenStruct` (added in
e1d9763c2, just over ten years ago!). The contract on the configuration
history object changed in e5556a2be. This commit brings the
`--no-history` option up to date with that refactor.