Test using Github Actions
We now test everything but 1.9 with GitHub Actions. GH Actions doesn't support
1.9 out of box and I am not sure how to write a spec that would use our custom
Docker image for 1.9 and the standard images for everything esle. That's why
CircleCI does not go away (for now, until we still support 1.9).

Luckily, I was able to get rid of the custom Docker image for 2.0 because GH
Actions stil support that version of Ruby.

Now the build code is much simpler and easier to follow. Looking forward to
getting rid of CircleCI completely!