Gemfile: run Rubocop only on Ruby 2.2.0
In https://github.com/pry/pry/pull/2212 I thought the build was fixed however
after mergin that PR with `master` showed that it wasn't the case.

Now I want to lock the Rubocop dependency to be installed only on Ruby 2.2.0 in
the hope that its dependencies doesn't clash with other gems. Still not sure
exactly what the problem is.