Weird method location shouldn't match unknown location
Methods that don't have a source location (e.g., C methods, or methods
created via metaprogramming or even `alias_method` to C methods) are not
reasonable possible matching methods for a "weird method" we need to
locate.

In this case `renamed_method_source_location` can also return `nil` if
the actual code in question is a bare script (i.e., no methods). If that
script is loaded via `eval` then we'll end up in the weird method path
in the first place, but no method matching can be found, and if a
no-source-location method exists, we'll return that.

Down the line that's particularly painful because the source loading
thinks it's a C method, but it can actually be from metaprogramming (and
`alias_method`!), and then `Pry::Method#pry_doc_info` raises an error,
and `wherami` breaks even though we already have valid `__FILE__` and
`__LINE__` values.